requirements.txt
altair==5.4.1
annotated-types==0.7.0
anyio==4.4.0
asgiref==3.8.1
asttokens==2.4.1
attrs==23.2.0
backcall==0.2.0
beautifulsoup4==4.12.3
bleach==6.1.0
blinker==1.8.2
cachetools==5.5.0
certifi==2024.2.2
chardet==3.0.4
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
comm==0.2.2
contourpy==1.2.1
coverage==7.6.8
cycler==0.12.1
debugpy==1.8.1
decorator==5.1.1
defusedxml==0.7.1
dj-database-url==2.3.0
Django==5.1.2
django-seed==0.3.1
djangorestframework==3.15.2
dnspython==2.6.1
docopt==0.6.2
email_validator==2.1.1
environ==1.0
executing==2.0.1
Faker==33.1.0
fastapi==0.111.0
fastapi-cli==0.0.4
fastjsonschema==2.19.1
filelock==3.13.4
filetype==1.2.0
fonttools==4.51.0
fsspec==2024.3.1
gitdb==4.0.11
GitPython==3.1.43
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
idna==2.7
imageio==2.35.1
intel-openmp==2021.4.0
ipykernel==6.29.4
ipython==8.12.3
jedi==0.19.1
Jinja2==3.1.3
joblib==1.4.2
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter_client==8.6.1
jupyter_core==5.7.2
jupyterlab_pygments==0.3.0
kiwisolver==1.4.5
lapx==0.5.10.post1
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.8.4
matplotlib-inline==0.1.6
mdurl==0.1.2
mistune==3.0.2
mkl==2021.4.0
mpmath==1.3.0
narwhals==1.8.1
nbclient==0.10.0
nbconvert==7.16.3
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.3
numpy==1.26.4
opencv-python==4.9.0.80
opencv-python-headless==4.10.0.84
orjson==3.10.5
packaging==24.0
pandas==2.2.2
pandocfilters==1.5.1
parso==0.8.4
pickleshare==0.7.5
pillow==10.3.0
pipreqs==0.5.0
platformdirs==4.2.0
postgis==1.0.4
prompt-toolkit==3.0.43
protobuf==5.28.1
psutil==5.9.8
psycopg2-binary==2.9.10
pure-eval==0.2.2
py-cpuinfo==9.0.0
pyarrow==17.0.0
pydantic==2.7.4
pydantic_core==2.18.4
pydeck==0.9.1
pygame==2.5.2
Pygments==2.17.2
PyMySQL==1.1.1
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
pytz==2024.1
PyYAML==6.0.1
pyzmq==25.1.2
referencing==0.34.0
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.7.1
roboflow==1.1.43
rpds-py==0.18.0
scikit-learn==1.6.0
scipy==1.13.0
seaborn==0.13.2
shellingham==1.5.4
sinesp_client==1.14
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.5
sqlparse==0.5.1
stack-data==0.6.3
starlette==0.37.2
streamlit==1.38.0
sympy==1.12
tbb==2021.12.0
tenacity==8.5.0
thop==0.1.1.post2209072238
thread==1.1.1
threadpoolctl==3.5.0
tinycss2==1.2.1
toml==0.10.2
toposort==1.10
torch==2.3.0
torchvision==0.18.0
tornado==6.4
tqdm==4.66.2
traitlets==5.14.2
typer==0.12.3
typing_extensions==4.11.0
tzdata==2024.1
ujson==5.10.0
ultralytics==8.3.27
ultralytics-thop==2.0.5
urllib3==1.24.3
uvicorn==0.30.1
watchdog==4.0.2
watchfiles==0.22.0
wcwidth==0.2.13
webencodings==0.5.1
websockets==12.0
yarg==0.1.9
